requirements.txt
asgiref==3.6.0
astroid==2.15.0
attrs==22.2.0
black==23.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.1
defusedxml==0.7.1
dill==0.3.6
dj-database-url==1.2.0
dj-rest-auth==3.0.0
Django==4.1.7
django-allauth==0.52.0
django-cors-headers==3.14.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-camel-case==1.4.2
gunicorn==20.1.0
idna==3.4
isort==5.12.0
jsonschema==4.17.3
lazy-object-proxy==1.9.0
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.7.0
oauthlib==3.2.2
packaging==23.0
pathspec==0.11.0
pipdeptree==2.5.2
platformdirs==3.1.0
psycopg2==2.9.5
pycparser==2.21
PyJWT==2.6.0
pylint==2.16.4
pylint-django==2.5.3
pylint-plugin-utils==0.7
pyright==1.1.298
pyrsistent==0.19.3
python3-openid==3.2.0
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
sqlparse==0.4.3
tomlkit==0.11.6
urllib3==1.26.15
whitenoise==6.3.0
wrapt==1.15.0
